Combine UserTest and User in all files
