+++ src/main/java/com/google/codeu/servlets/FormHandlerServlet.java
@@ -1,4 +1,4 @@
-package io.happycoding.servlets;
+package com.google.codeu.servlets;
 
 import com.google.appengine.api.blobstore.BlobInfo;
 import com.google.appengine.api.blobstore.BlobInfoFactory;
