+++ src/main/java/com/google/codeu/servlets/SkillSearchServlet.java
@@ -1,9 +1,8 @@
 package com.google.codeu.servlets;
 
 import com.google.codeu.data.Datastore;
-import com.google.codeu.data.UserTest;
+import com.google.codeu.data.User;
 import com.google.gson.Gson;
-import jdk.nashorn.internal.objects.NativeJSON;
 
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -36,16 +35,16 @@ public class SkillSearchServlet extends HttpServlet {
         String paramUsers = request.getParameter("testUsers");
         System.out.println(paramUsers);
         Gson gson = new Gson();
-        Type type = new TypeToken<ArrayList<UserTest>>() {
+        Type type = new TypeToken<ArrayList<User>>() {
         }.getType();
-        ArrayList<UserTest> arr = gson.fromJson(paramUsers, type);
+        ArrayList<User> arr = gson.fromJson(paramUsers, type);
 
         //all users
-        ArrayList<UserTest> users = new ArrayList<UserTest>(arr.size());
+        ArrayList<User> users = new ArrayList<User>(arr.size());
 
         for (int i = 0; i < arr.size(); i++) {
             HashMap<String, String> info = new HashMap<>();
-            UserTest user = arr.get(i);
+            User user = arr.get(i);
             users.add(user);
         }
 
@@ -53,10 +52,10 @@ public class SkillSearchServlet extends HttpServlet {
         String[] categories = {"Design", "Photo", "Sports", "Music", "Technology", "Language", "Culinary"};
 
         //users categorized by learnCategory
-        HashMap<String, ArrayList<UserTest>> cat_users = categorizeUsers(users, categories);
+        HashMap<String, ArrayList<User>> cat_users = categorizeUsers(users, categories);
 
         //select targetskill users
-        ArrayList<UserTest> selected = cat_users.get(targetSkill);
+        ArrayList<User> selected = cat_users.get(targetSkill);
 
         response.setContentType("application/json");
 
@@ -72,13 +71,13 @@ public class SkillSearchServlet extends HttpServlet {
     }
 
     //categorize users
-    private HashMap<String, ArrayList<UserTest>> categorizeUsers(ArrayList<UserTest> users,
+    private HashMap<String, ArrayList<User>> categorizeUsers(ArrayList<User> users,
                                                                  String[] categories) {
-        HashMap<String, ArrayList<UserTest>> cat_users = new HashMap<>();
+        HashMap<String, ArrayList<User>> cat_users = new HashMap<>();
         users.forEach((user) -> {
             String skill = user.getTeachCategory();
             if (cat_users.get(skill) == null) {
-                ArrayList<UserTest> userList = new ArrayList<>();
+                ArrayList<User> userList = new ArrayList<>();
                 userList.add(user);
                 cat_users.put(skill, userList);
             } else {
